fix(get_branched_repo): handle OSS/enterpise releases
code was hardcode incorrectly, and was working only for master
and not for enterprise, nor releases

all the function releated to this function where moved
to `sdcm.utils.version_utils` to avoid cyclic dependecy.
